requirements.txt
awesome-slugify==1.6.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bs4==0.0.1
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-autocomplete-light==3.1.5
django-autoslug==1.9.3
django-cleanup==0.4.2
django-coverage==1.2.4
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-model-utils==2.5
django-nose==1.4.3
django-passwords==0.3.10
django-simple-captcha==0.5.1
django-suit==0.2.18
django-taggit==0.18.1
djorm-pgarray==1.2
factory-boy==2.7.0
fake-factory==0.5.7
gunicorn==19.4.5
invoke==0.12.2
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
nose==1.3.7
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.4
regex==2016.4.25
selenium==2.53.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.19
static3==0.6.1
traitlets==4.2.1
Unidecode==0.4.19
Unipath==1.1
Werkzeug==0.11.9
whitenoise==3.0
